Added acccess control - still debugging

+++ admin/views/jtraxes/view.html.php
@@ -15,6 +15,7 @@ use Joomla\CMS\MVC\View\HtmlView as BaseHtmlView;
 use Joomla\CMS\Toolbar\ToolbarHelper;
 use Joomla\CMS\Language\Text;
 use Joomla\CMS\Factory;
+use Joomla\CMS\Helper\ContentHelper;
 
 /**
  * View class for a list of tracking codes.
@@ -32,7 +33,8 @@ class JtraxViewJtraxes extends BaseHtmlView
      */
     public function display($tpl = null)
     {
-        $this->items		 = $this->get('Items');
+		$this->canDo		 = ContentHelper::getActions('com_jtrax');
+		$this->items		 = $this->get('Items');
         $this->pagination	 = $this->get('Pagination');
         $this->state		 = $this->get('State');
         $this->filterForm	 = $this->get('FilterForm');
@@ -58,13 +60,30 @@ class JtraxViewJtraxes extends BaseHtmlView
     protected function addToolbar()
     {
         ToolbarHelper::title(Text::_('COM_JTRAX_TITLE_MAIN'), 'stack');
-
-        ToolbarHelper::addNew('jtrax.add');
-        ToolbarHelper::editList('jtrax.edit');
-        ToolbarHelper::publish('jtraxes.publish', 'JTOOLBAR_PUBLISH', true);
-        ToolbarHelper::unpublish('jtraxes.unpublish', 'JTOOLBAR_UNPUBLISH', true);
-        ToolbarHelper::deleteList(Text::_('COM_JTRAX_DELETE_QUESTION'), 'jtraxes.delete', 'JTOOLBAR_DELETE');
-
-        ToolbarHelper::preferences('com_jtrax');
+			if ($this->canDo->get('core.create')) 
+			{
+				ToolbarHelper::addNew('jtrax.add');
+			}
+			if ($this->canDo->get('core.edit'))
+			{
+				ToolbarHelper::editList('jtrax.edit');
+			}
+			if ($this->canDo->get('core.edit.state'))
+			{
+				ToolbarHelper::publish('jtraxes.publish', 'JTOOLBAR_PUBLISH', true);
+			}
+			if ($this->canDo->get('core.edit.state'))
+			{
+				ToolbarHelper::unpublish('jtraxes.unpublish', 'JTOOLBAR_UNPUBLISH', true);
+			}
+			if ($this->canDo->get('core.delete')) 
+			{
+			ToolbarHelper::deleteList(Text::_('COM_JTRAX_DELETE_QUESTION'), 'jtraxes.delete', 'JTOOLBAR_DELETE');
+			}
+		
+		if ($this->canDo->get('core.admin'))
+		{
+			ToolbarHelper::preferences('com_jtrax');
+		}
     }
 }
